+++ Assets/Neoxider/Scripts/Cards/Core/Enums/CardLocation.cs
@@ -15,3 +15,4 @@ namespace Neo.Cards
 
 
 
+