+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcAggroFollowCore.cs
@@ -70,3 +70,4 @@ namespace Neo.NPC.Navigation
 
 
 
+