+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcAnimationCore.cs
@@ -52,3 +52,4 @@ namespace Neo.NPC.Navigation
 
 
 
+