+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcDestinationResolver.cs
@@ -59,3 +59,4 @@ namespace Neo.NPC.Navigation
 
 
 
+