+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcFollowTargetCore.cs
@@ -246,3 +246,4 @@ namespace Neo.NPC.Navigation
 
 
 
+