+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcNavAgentCore.cs
@@ -89,3 +89,4 @@ namespace Neo.NPC.Navigation
 
 
 
+