+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcPatrolCore.cs
@@ -283,3 +283,4 @@ namespace Neo.NPC.Navigation
 
 
 
+